% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_enrollment_df}
\alias{create_empty_enrollment_df}
\title{Create empty enrollment data frame}
\usage{
create_empty_enrollment_df(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{"Campus", "District", or "State"}
}
\value{
Empty data frame with expected columns
}
\description{
Create empty enrollment data frame
}
\keyword{internal}
