% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_empty_nhdoe_df}
\alias{create_empty_nhdoe_df}
\title{Create empty NH DOE data frame}
\usage{
create_empty_nhdoe_df(level = "school")
}
\arguments{
\item{level}{"school" or "district"}
}
\value{
Empty data frame with expected columns
}
\description{
Creates an empty data frame with expected NH DOE column structure.
}
\keyword{internal}
