% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_iplatform_report}
\alias{download_iplatform_report}
\title{Download report from NH DOE iPlatform}
\usage{
download_iplatform_report(report_base, report_path, end_year, level)
}
\arguments{
\item{report_base}{Base URL for reports}

\item{report_path}{Path to specific report}

\item{end_year}{School year end}

\item{level}{"district" or "school"}
}
\value{
Data frame with enrollment data
}
\description{
Attempts to download a report from the NH DOE iPlatform SSRS system.
}
\keyword{internal}
