% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nhdoe_district_enrollment}
\alias{download_nhdoe_district_enrollment}
\title{Download district enrollment data from NH DOE iPlatform}
\usage{
download_nhdoe_district_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with district enrollment data
}
\description{
Downloads the District Fall Enrollment report from NH DOE iPlatform.
The report includes enrollment by grade level for all districts.
}
\keyword{internal}
