% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nhdoe_school_enrollment}
\alias{download_nhdoe_school_enrollment}
\title{Download school enrollment data from NH DOE iPlatform}
\usage{
download_nhdoe_school_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with school enrollment data
}
\description{
Downloads the Public School Enrollments by Grade report from NH DOE iPlatform.
}
\keyword{internal}
