% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nhdoe_static_enrollment}
\alias{download_nhdoe_static_enrollment}
\title{Download NH DOE static enrollment files}
\usage{
download_nhdoe_static_enrollment(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"district" or "school"}
}
\value{
Data frame with enrollment data, or NULL if not available
}
\description{
Attempts to download enrollment data from static files that NH DOE
may publish on their website.
}
\keyword{internal}
