% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_district_id}
\alias{parse_district_id}
\title{Parse NH district ID}
\usage{
parse_district_id(x)
}
\arguments{
\item{x}{Character vector of district IDs}
}
\value{
Cleaned district ID
}
\description{
NH uses a district numbering system. Districts are identified by
SAU (School Administrative Unit) number and district code.
}
\keyword{internal}
