% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_year}
\alias{validate_year}
\title{Validate year parameter}
\usage{
validate_year(end_year)
}
\arguments{
\item{end_year}{Year to validate}
}
\value{
TRUE if valid, throws error if not
}
\description{
Validate year parameter
}
\keyword{internal}
